+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Added paste support to venn. Adjust size default size of trees.
 - Venn lists can be removed.
 - Venn now supports heatmaps. Rendering moved into hook rather than page. Support for gaps in heatmaps.
 - Ability to edit individual axes.